[VisionGlass] New design of the PhoneUI
This PR moves all the style bits of the visionglass_layout.xml into
the themes.xml resource file. This way we can define a more structured
and flexible design.

Additionally, this change provides a new design of the PhoneUI to
provide a more android-friendly UI, with rounded-corner elements and
an homogeneous color pattern.