Add audio feedback for keyboard presses
When typing on the keyboard, there is no sound that confirms selection. Without sound
cues, users are forced to rely heavily on visual confirmation to ensure their inputs
have been registered, often causing them to continuously check the input field.

We're adding audio feedback to insert (letters) and the backspace keys. Special keys
are not handled now but could be in the future.