[VisionGlass] Disable some PhoneUI buttons in WebXR sessions
During Immersive sessions we would need to avoid clicks on some of
the widgets provided by the PhoneUI.

This PR introduces a new BindingAdapter "dimmingEnabled" to apply a
dimming visual effect on a widget and disable its input methods.

Additionally, the mIsPresentingImmersive field of the VRBBrowserActivity
class has become a LiveMutable object, so that we can observe its changes
and trigger the dimming actions in the PlatformActivity, if implemented.

Finally, the PR adds the new bindingAdapter in the corresponding
elements of the visionglass_layout.xml; specifically, we are going to
disable all the UI elements, but the Touchpad and the Back button.