[VisionGlass] Disable the PhoneUI buttons when window is fullscreen
The PR adds a new method in the WindowListener interface to notify that
the WindowViewModel's "isFullscreen" field changes. The VRBrowserActiviy
class, as implemented of the Windows' Delegate interface, handles the
event and updates the PlatformActivityPlugin.

This change adds a new MutableLive attribute "mIsFullscreen" to the
PhoneUIViewModel, which is managed by the PlatformActivityPlugin.

The corresponding buttons in the visionglass_layout.xml are going to
be dimmed based on the value of this "mIsFullscreen" live attribute.