[VisionGlass] Restart activity when unplugging the glasses (#1795)
Then the glasses are desconnected the Display of the Presentation
is destroyed.  Creating another Presentation is mandatory on
reconnecting because the Display would be different. This means
the native machinery is no longer valid. The RenderContext is
using a different thread than the one that is using the new
GLSurfaceView of the new Presentation.

We'd have to recreate the whole BrowserWorld. Unfortunately
this means that we'd have to re-run all the application creation
Java code because among other things, we create the Widgets
there. As this is all very coupled, there is no easy way to do
that outside the application creation code flow.

This means that the only way to get Wolvic running on disconnection
is basically to restart the activity. The process is a bit slow
because we need to initialize everything again but it's safe
and we also ensure that the state is saved and recreated before
restarting.

Fixes #1764