Create the fullscreen observer after the listeners array initialization
The commit 406a611 cause a crash due to the WindowWidget's mListeners
attribute was not initialized the fist time the onIsFullscreeen
callback.

We were creating an observer for the isFullScreen MutalbleLive field
and the onIsFullscreen is the callback of such observer. When the
observer is created, the callback is called with the current value
of the MutableLive object. We were creating the observer before the
mListeners field was initialized.

This PR fixes the crash by moving the observer creation after
completely the mListeners initialization.