Use hover to focus windows
When in multiwindow mode, users need to click on the browser windows
to explicitly set them as active and thus, move widgets like the
URL bar or the navigation bar next to the selected window.

In order to improve it, we can do that automatically just by
hovering the window. This will be the new default as it looks
like it's a better behaviour in general, at least more
comfortable. However, as it's a breaking change, we are adding
a setting to the controller settings to revert to the previous
behaviour of click-to-select.