Adjust navigation taking New Tab into account
Use WindowWidget.onLocationChange() to decide whether the New Tab,
the Library, or Web content should be shown.

By passing the responsibility of keeping track of navigation to
the session, we are able to remove many of the bugs caused by
the interaction between the New Tab and the Web content.

The library is not part of that navigation, which was already
the case before.

Rename WindowWidget.showLibraryPanel() and hidePanel()
to showLibrary() and closeLibrary() for clarity.

Remove explicit calls to hide the Library, since these will be
managed internally by WindowWidget.

Remove unneeded fields in the WindowViewModel, since we will
use the session to keep track of back and forward navigation.