Remote experiences in New Tab
Download the list of remote experiences in JSON format and display
it in the New Tab view.

The list is based on the content of out existing homepage, with
some small changes to make it easier to process and display.
The process is very similar to how props.json is managed.

This change also adds a utility class to download and display
remote images asynchronously.