Add announcements to New Tab
Add an announcements feature that receives news items from a remote JSON file.
Announcements are defined in the `ANNOUNCEMENTS_ENDPOINT` URL.

The process to fetch and present them is similar to what we do with remote
experiences and similar content.

Each announcement includes a title and body text, and optionally an image and
a link. Text fields support multiple languages.