Clean up some issues when changing the homepage in the Settings
Clean up the code and fix some edge cases where the homepage was
being incorreclty updated.

Use the current homepage value for the URL edit field, because
otherwise it was far too confusing.

Do not assume that our home webpage will be the default (for example,
the default might be to show the native New Tab UI).