Dismiss announcements on New Tab
Add onDismissed to AnnouncementsAdapter.ClickListener and call it when the user
clicks on the "close" button on the announcement.

SettingsStore keeps a list of dismissed announcements so we can avoid showing
those that the user has already closed.

SettingsViewModel.visibleAnnouncements contains a filtered list of announcements,
making it easier to keep the UI up to date.

Update layout of announcement item.