Fix resource leaks when switching between tabs
Several fixes to the process of switching between different tabs,
specially when they are showing a native UI line the New Tab.

When using layers, ensure that the surface held by WindowWidget
is released at the same time as the session is notified that
it is no longer valid.

Move setupListeners() towards the end of the setSession() because
it will cause session listeners to be called synchronously with
the new values and we need to have a consistent state at that point.

Ensure that listeners are not set up on a session that is not the
current one.

Fixes https://github.com/Igalia/wolvic/issues/1841