Fix resource leaks when switching between tabs
Fix a race condition that creates a resource leak which ends up
crashing the application.

When we switch between two tabs showing the New Tab we first request
a new Surface for the new session, followed by a Surface for the
native UI.

However, both Surfaces might arrive once the native UI is already in
place. In that case, both Surfaces will be applied to the native UI
while the browser session keeps a reference to the Surface that
belonged to the previous session, which might cause errors.

To fix this, when displaying the native UI we will clear mSurface
after pausing the compositor.

Additionally, we move setupListeners() towards the end of
setSession() because it will cause session listeners to be called
synchronously with the new values and we need to have a consistent
state at that point.

Fixes https://github.com/Igalia/wolvic/issues/1841