Check if layers are enabled when deciding to update the widget surface
When deciding to update the widget surface, the isLayer()
method is unreliable because it can return false even if
layers are enabled, for example if mSurface becomes null.
With this change, we will use the settings for this check.

This fixes the bug that the Library is not shown with Chromium.