Chromium uses "chrome://newtab/" as New Tab URL
When using Chromium, the New Tab UI will be shown when the browser
navigates to "chrome://newtab/" which is Chromium's default empty
start page.

The reason for this change is that Chrome returns an error when we
use the "about://newtab" URL and loads "chrome://newtab/" instead.

Fixes https://github.com/Igalia/wolvic/issues/1868