Set the pressed status for the Settings button in the tray
The settings button in the tray was never show as pressed. That's because
we didn't have the data in the view model to activate/deactivate it.

We added a visibility changed interface to the UIDialog class which is
called whenever show()/hide() are called. We can use that information
later to fillin a boolean in the view model that will be used later
to show the pressed/unpressed status of the button.