Migrate WSession::getSessionFinder to an asynchronous version
This is required to get the find-in-page feature working in Chromium. In the case of this
engine the session finder is not always available because sometimes we need to wait a bit
until the WebContents object is available. That's why we need to move to an asynchronous
version of this method as a first step.