Keyboard layout update
Replace the vertical group of control buttons so each button
may be positioned separately. This also required changes in
hover management.

Use a slightly larger and more visible theming for the voice
and close buttons.

Replace the "keyboardMoveButton" with a bar like the one that
we use to move windows.

Update keyboard size calculations accordingly.