Added a delay to hover to focus
We recently added the possibility of switching the focus between windows
by just hovering them. The problem is that switching everytime the window
gets a focus is problematic because there are too many false positives.
Users move their hands/controllers "erratically" so the focus might easily
switch to another window when the users don't really want to.

By adding a 500msecs delay we'll remove most of them. From now on it's
even possible to quickly hover windows alternatively and then focus won't
change.