Use the orientation of the focused controller to move windows
This was the original intention but instead of checking the focused
property of the controllers we're using enabled. If both hands or
controllers were used the the first one (the right one) was always
selected because it was the first enabled one in the list (but not
the only one).

Fixes #1885