Show a warning dialog when disabling hand tracking
The hand tracking setting is ephemeral, hand tracking is automatically restored
on the next app launch to prevent issues with unavailable controllers. Users
have been confused because indeed it looks like that their preferences are not
saved, but that's done on purpouse. We must inform them about the volatility of
the setting.

From now on a dialog is shown when hand tracking is disabled.