Show the ETP widget when clicking on the Advanced button
The current code was always showing the Password Exceptions dialog instead.
The problem was that we had a switch to select the proper dialog and the
ETP option didn't have a break at the end so it was falling back to the next
option which was the password exceptions option.

Fixes #1901