Securize refresh rate selection
If the XR_FB_display_refresh_rate extension was available we assumed
that the API will always return a non empty list of supported refresh
rates. However in some particular configurations the extension might
appear as available but no refresh rates will be ever returned. That
could cause crashes due to assertions.

As this is an extension and not a critical piece of the main execution
path we should just report the error and bail out.

This was happening to me when trying to run the AOSP flavour in a
Quest device. With this fix Wolvic can be run in Quest devices even
with the AOSP flavour.