Use a custom annotation interface for scroll direction
It's safer and improves the understandability of the code. The new
definitions replaced the old DEFAULT value by NATURAL which matches
what we show in the UI. There is still the need for an invalid value
(-1 in this case) because we apply caching and we need to know
whether the value has been initialized.