Move layer support detection to the backend
The check for support for composition layers must be done in the
backend (OpenXR) which is the one with enough data to decide.

We were doing it the other way around, checking which devices did
support them and forcing it in the Java side because we knew that
would work on the native side.

It's much safer and less error prone to do it on the backend. This
forces us to make some Java calls asynchronous because we don't
know whether the backend has done its checks or not.