Improvements and fixes for the build system
First of all we're specifying that the gecko substitution should be
done for the omni version of the GeckoView packages. By default if
omni is not specified then gradle expects the lite version. For a
full fledged browser the omni version should be used.

That brings some challenges. For example glean is provided both
by GV and android components so we had to had some dependency
analysis code to select the one from GV (this is also done by
Mozilla in their reference browser implementation).

Last but not least, the dependency analysis forced us to bump
the gradle and gradle plugin versions.