Revamp the computation of controllers' poses
We used to retrieve both aim and grip poses using the local space as
base space. That's problematic because WebXR for example expects the aim
to be based on the grip pose instead. The code is now simpler as we
don't need to do special things for Chromium. This also allowed us to
remove some workarounds like the artificial offsets we had to add to the
position of the controllers or correction angles to fix the orientation
of some poses.

Also the current code was positioning the controller using the aim when
it should really use the grip. We used to do that because there are
cases where there is no grip (like when using hands) but for those cases
we can always fallback to the aim pose. This fixes the weird
orientations shown by controllers when rotating them, they used to
rotate around the bottom of the controller instead of using the actual
rotation center of the movement.

Another important change is that we now properly separate the
computation of controller transform, beam transform and immersive beam
transform depending on the rendering mode. The reason for that is clear,
WebXR expects all the poses in local space. For the distance to floor it
uses the SittingToStandingTransform. For the 2D browsing case however,
we need to include the distance to the floor in the poses. Also some of
those transforms are useless for some rendering modes, for example WebXR
does not need the beam transform, and 2D rendering does not need the
immersive beam.Revamp the computation of controllers' poses.

HVR backend is a bit special as it does not compute correctly poses if
the base space is not the local space. That forced us to include several
ifdefs for the HVR special behaviour.

Last but not least, we had to modify the 3D controller models to remove
some extra rotations we had to add in the past to better align the
position of the virtual controllers with the position of the real ones.

Fixes #1451, #1832

WIP