Check pose's valid bit instead of tracked bit
It's safer, more reliable and error-resistant to check the valid bit
than the tracked bit. First of all, from the functionality POV we do
really want to check that the pose is valid, it does not really matter
whether it's tracked or emulated by the runtime. Also many runtimes are
buggy and do not properly fill in the tracked bit. Last but not least it
allows us to remove fallbacks for those buggy runtimes.