[Chromium] Fix the total matches count in find in page
The code assumes that the Web Engine reports the number of matches only
once (as Gecko does). However Chromium reports partial results on the go.
As the code was only considering the first notification from the engine
we were not properly picking the right values for current and total
number of matches (just the first partial results).

Fixes #1481